feat(core): enforce allowed object types for tree operations

+++ stustapay/core/service/user.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 
 from stustapay.core.config import Config
 from stustapay.core.schema.terminal import Terminal
-from stustapay.core.schema.tree import Node
+from stustapay.core.schema.tree import Node, ObjectType
 from stustapay.core.schema.user import (
     CurrentUser,
     NewUser,
@@ -67,9 +67,8 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user_role])
     async def create_user_role(self, *, conn: Connection, node: Node, new_role: NewUserRole) -> UserRole:
-        # TODO: TREE visibility
         role_id = await conn.fetchval(
             "insert into user_role (node_id, name, is_privileged) values ($1, $2, $3) returning id",
             node.id,
@@ -88,14 +87,13 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user_role])
     async def update_user_role_privileges(
-        self, *, conn: Connection, role_id: int, is_privileged: bool, privileges: list[Privilege]
+        self, *, conn: Connection, node: Node, role_id: int, is_privileged: bool, privileges: list[Privilege]
     ) -> UserRole:
-        # TODO: TREE visibility
         role = await _get_user_role(conn=conn, role_id=role_id)
-        if role is None:
-            raise NotFound(element_typ="user_role", element_id=str(role))
+        if role is None or role.node_id not in node.ids_to_root:
+            raise NotFound(element_typ="user_role", element_id=role_id)
 
         await conn.execute("update user_role set is_privileged = $2 where id = $1", role_id, is_privileged)
 
@@ -111,12 +109,10 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
-    async def delete_user_role(self, *, conn: Connection, role_id: int) -> bool:
-        # TODO: TREE visibility
+    @requires_node(object_types=[ObjectType.user_role])
+    async def delete_user_role(self, *, conn: Connection, node: Node, role_id: int) -> bool:
         result = await conn.execute(
-            "delete from user_role where id = $1",
-            role_id,
+            "delete from user_role where id = $1 and node_id = any($2)", role_id, node.ids_to_root
         )
         return result != "DELETE 0"
 
@@ -124,12 +120,11 @@ class UserService(DBService):
     async def _update_user_roles(
         *, conn: Connection, node: Node, user_id: int, role_names: list[str], delete_before_insert=False
     ):
-        # TODO: TREE visibility
+        """The caller of this function needs to make sure that the update is valid for the given user and node."""
         if delete_before_insert:
             await conn.execute("delete from user_to_role where user_id = $1", user_id)
 
         for role_name in role_names:
-            # TODO: NODE_ID incorporate node_id
             role_id = await conn.fetchval(
                 "select id from user_role where name = $1 and node_id = any($2)", role_name, node.ids_to_root
             )
@@ -192,7 +187,6 @@ class UserService(DBService):
         new_user: NewUser,
         password: Optional[str] = None,
     ) -> User:
-        # TODO: TREE visibility
         node = await fetch_node(conn=conn, node_id=node_id)
         assert node is not None
         return await self._create_user(
@@ -201,7 +195,7 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user])
     async def create_user(
         self,
         *,
@@ -211,15 +205,16 @@ class UserService(DBService):
         new_user: NewUser,
         password: Optional[str] = None,
     ) -> User:
-        # TODO: TREE visibility
         return await self._create_user(
             conn=conn, creating_user_id=current_user.id, node=node, new_user=new_user, password=password
         )
 
     @staticmethod
-    async def _contains_privileged_roles(conn: Connection, role_names: list[str]) -> bool:
+    async def _contains_privileged_roles(conn: Connection, node: Node, role_names: list[str]) -> bool:
         res = await conn.fetchval(
-            "select true from user_role where name = any($1::text array) and is_privileged", role_names
+            "select true from user_role where name = any($1::text array) and is_privileged and node_id = any($2)",
+            role_names,
+            node.ids_to_root,
         )
         return res is not None
 
@@ -227,7 +222,9 @@ class UserService(DBService):
     @requires_terminal([Privilege.user_management])
     async def create_user_terminal(self, *, conn: Connection, current_user: CurrentUser, new_user: NewUser) -> User:
         # TODO: TREE visibility
-        if await self._contains_privileged_roles(conn=conn, role_names=new_user.role_names):
+        node = await fetch_node(conn=conn, node_id=current_user.node_id)
+        assert node is not None
+        if await self._contains_privileged_roles(conn=conn, node=node, role_names=new_user.role_names):
             raise AccessDenied("Cannot promote users to privileged roles on a terminal")
 
         return await self.create_user_with_tag(  # pylint: disable=missing-kwoa
@@ -237,10 +234,17 @@ class UserService(DBService):
     @with_db_transaction
     @requires_terminal([Privilege.user_management])
     async def update_user_roles_terminal(
-        self, *, conn: Connection, current_terminal: Terminal, user_tag_uid: int, role_names: list[str]
+        self,
+        *,
+        conn: Connection,
+        current_terminal: Terminal,
+        current_user: CurrentUser,
+        user_tag_uid: int,
+        role_names: list[str],
     ) -> User:
-        # TODO: TREE visibility
-        if await self._contains_privileged_roles(conn=conn, role_names=role_names):
+        node = await fetch_node(conn=conn, node_id=current_user.node_id)
+        assert node is not None
+        if await self._contains_privileged_roles(conn=conn, node=node, role_names=role_names):
             raise AccessDenied("Cannot promote users to privileged roles on a terminal")
 
         node = await fetch_node(conn=conn, node_id=current_terminal.till.node_id)
@@ -258,10 +262,8 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
-    async def create_user_with_tag(
-        self, *, conn: Connection, node: Node, current_user: CurrentUser, new_user: NewUser
-    ) -> User:
+    @requires_node(object_types=[ObjectType.user])
+    async def create_user_with_tag(self, *, conn: Connection, current_user: CurrentUser, new_user: NewUser) -> User:
         # TODO: TREE visibility
         """
         Create a user at a Terminal, where a name and the user tag must be provided
@@ -269,6 +271,8 @@ class UserService(DBService):
 
         returns the created user
         """
+        node = await fetch_node(conn=conn, node_id=current_user.node_id)
+        assert node is not None
         user_tag_uid = await conn.fetchval("select uid from user_tag where uid = $1", new_user.user_tag_uid)
         if user_tag_uid is None:
             raise InvalidArgument(f"Tag uid {format_user_tag_uid(new_user.user_tag_uid)} not found")
@@ -294,12 +298,11 @@ class UserService(DBService):
 
     @staticmethod
     async def _get_user(*, conn: Connection, node: Node, user_id: int) -> User:
-        # TODO: TREE visibility
         user = await conn.fetch_maybe_one(
             User, "select * from user_with_roles where id = $1 and node_id = any($2)", user_id, node.ids_to_root
         )
         if user is None:
-            raise NotFound(element_typ="user", element_id=str(user_id))
+            raise NotFound(element_typ="user", element_id=user_id)
         return user
 
     @with_db_transaction
@@ -312,12 +315,13 @@ class UserService(DBService):
         row = await conn.fetchrow(
             "update usr "
             "set login = $2, description = $3, display_name = $4, user_tag_uid = $5 "
-            "where id = $1 returning id",
+            "where id = $1 and node_id = any($6) returning id",
             user_id,
             user.login,
             user.description,
             user.display_name,
             user.user_tag_uid,
+            node.ids_to_root,
         )
         if row is None:
             raise NotFound(element_typ="user", element_id=str(user_id))
@@ -330,7 +334,7 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user])
     async def update_user_roles(
         self, *, conn: Connection, node: Node, user_id: int, role_names: list[str]
     ) -> Optional[User]:
@@ -349,14 +353,14 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user])
     async def update_user(self, *, conn: Connection, node: Node, user_id: int, user: UserWithoutId) -> Optional[User]:
         # TODO: TREE visibility
         return await self._update_user(conn=conn, node=node, user_id=user_id, user=user)
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user])
     async def change_user_password(
         self, *, conn: Connection, node: Node, user_id: int, new_password: str
     ) -> Optional[User]:
@@ -374,7 +378,7 @@ class UserService(DBService):
 
     @with_db_transaction
     @requires_user([Privilege.user_management])
-    @requires_node()
+    @requires_node(object_types=[ObjectType.user])
     async def delete_user(self, *, conn: Connection, user_id: int) -> bool:
         # TODO: TREE visibility
         result = await conn.execute(
